Sample Template Project – User Profile Setup
This is a sample template project that you could open up in the Template Tool software and explore, edit to suit your needs, and/or use to compile questions about to send to me at info@templatetool.io.  I will create and link to a YouTube video tutorial shortly that will show you how to explore and edit this sample file.